Created a temporary folder for RISC-V testing
