+++ test/riscv/src/intercept_sys_fcntl.c
@@ -30,9 +30,29 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include "libsyscall_intercept_hook_point.h"
+#include <syscall.h>
+#include <unistd.h>
 #include <fcntl.h>
 
-int main() {
-    int fd = openat(AT_FDCWD, "intercepting_original_file.txt", O_RDONLY);
-    int fd_dup = fcntl(fd, F_DUPFD, 0);
+static int
+hook(long syscall_number,
+     long arg0, long arg1,
+     long arg2, long arg3,
+     long arg4, long arg5,
+     long *result)
+{
+    if (syscall_number == SYS_fcntl) {
+        int fd = openat(AT_FDCWD, "../testfile.txt", O_RDWR | O_CREAT | O_TRUNC, 0666);
+        int ret = syscall_no_intercept(syscall_number, fd, arg1, arg2, arg3, arg4, arg5);
+        *result = ret;
+        return 0;
+    }
+    return 1;
+}
+
+static __attribute__((constructor)) void
+init(void)
+{
+    intercept_hook_point = hook;
 }
