+++ test/riscv/src/intercept_sys_openat.c
