+++ test/riscv/src/intercept_sys_statx.c
