+++ test/riscv/src/statx_test.c
