+++ test/riscv/src/write_test.c
